(optimize) move Badge (previously NumberBadge) to BadgedButton since it's only used in it, and delete the component file
